Implement screens navigation and add shoe item logic

+++ app/src/main/java/com/tunde/shoeinventory/model/Shoe.kt
@@ -2,7 +2,7 @@ package com.tunde.shoeinventory.model
 
 data class Shoe(
     var name: String,
-    var size: Int,
+    var size: Double,
     var company: String,
     var color: String,
     var description: String
